Time: 1 ms (71.54%) | Memory: 40.8 MB (59.90%) - LeetSync

+++ 29-divide-two-integers/divide-two-integers.java
@@ -3,7 +3,7 @@ class Solution {
         if(divisor ==1 ) return dividend;
         if(dividend==divisor) return 1;
         long ans = 0;
-        long sign = 1;
+        int sign = 1;
         if((dividend>=0 && divisor<0) || (dividend <0 && divisor>=0))
             sign = -1;
         long n = Math.abs((long)dividend);
